refactor: move Users features to domain folder
