+++ src/pages/signup.tsx
@@ -14,7 +14,7 @@ import {
   CustomerSignupForm,
   ErrorNotification,
   SuccessNotification,
-} from "../features/CustomerSignupPage";
+} from "../domains/users/CustomerSignupPage";
 
 const StyledDiv = styled.div`
   @import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,300;0,400;0,500;0,600;0,700;1,400&display=swap");
